requirements.txt
catboost==1.2.7
fastapi==0.115.8
joblib==1.3.2
numpy==1.24.4
pandas==2.2.3
polars==1.22.0
pyarrow==17.0.0
pydantic==2.10.6
PyYAML==6.0.1
scikit_learn==1.6.1
scipy==1.15.1
uvicorn==0.34.0
xgboost==2.1.4
